Prepared statement適用中 clean up実施
git-svn-id: file:///Users/hiro/repository/beauties@158 b1584ab8-daee-4b4e-b8c6-5a3b018d5d45

+++ Beauties/src/beauties/record/view/dialog/CompositeMove.java
@@ -203,7 +203,7 @@ class CompositeMove extends Composite {
 		mBookFromCombo.select(mBookIdList.indexOf(mExpenseRecord.getBookId()));
 		mBookToCombo.select(mBookIdList.indexOf(mIncomeRecord.getBookId()));
 
-		mValueSpinner.setSelection((int) mIncomeRecord.getIncome());
+		mValueSpinner.setSelection(mIncomeRecord.getIncome());
 		mFrequencySpinner.setSelection(mIncomeRecord.getFrequency());
 
 		if (!"".equals(mIncomeRecord.getNote()))
